Criado script utilitário para Cooldown
O script ImmunityAfterTakeDamage agora é um tipo de cooldown
