+++ Assets/Scripts/Utility/Cooldown.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55f5451e108eea10c941693c34e1e70a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
